refactor: move ButtonGroup into SharedComponents folder
