+++ src/components/BioCardList/BioCardList.tsx
@@ -4,10 +4,10 @@ import BioModal from "../BioModal";
 import { BioData } from "../BioModal/BioModal";
 import { useBioModal } from "../BioModal/useBioModal";
 import Grid, { GridArea } from "../Grid";
-import Flex, { FlexProps } from "../SharedComponents/Flex";
 
 import BioCardListItem from "./BioCardListItem";
 
+import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 import Box from "@/components/SharedComponents/Box";
 
 type BioCardListProps = FlexProps & {
