+++ src/components/BrowserWidthBar/BrowserWidthBar.tsx
@@ -2,7 +2,8 @@ import { FC } from "react";
 import styled from "styled-components";
 
 import background from "../../styles/utils/background";
-import Flex, { FlexProps } from "../SharedComponents/Flex";
+
+import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 
 const Root = styled(Flex)`
   position: relative;
