+++ src/components/Card/SubjectUnitLessonTitleCard.tsx
@@ -1,10 +1,10 @@
 import { FC } from "react";
 
-import Flex, { FlexProps } from "../SharedComponents/Flex";
 import { Heading } from "../Typography";
 import BoxBorders from "../SpriteSheet/BrushSvgs/BoxBorders/BoxBorders";
 import SubjectIcon from "../SubjectIcon";
 
+import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 import Box from "@/components/SharedComponents/Box";
 
 export const titleCardIconBackground = {
