+++ src/components/Circle/Circle.tsx
@@ -2,7 +2,8 @@ import { FC } from "react";
 
 import { PixelSpacing } from "../../styles/theme";
 import { ResponsiveValues } from "../../styles/utils/responsive";
-import Flex, { FlexProps } from "../SharedComponents/Flex";
+
+import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 
 export type CircleProps = FlexProps & {
   children?: React.ReactNode;
