+++ src/components/DropdownSelect/Select.tsx
@@ -5,7 +5,6 @@ import { useObjectRef } from "@react-aria/utils";
 import { useSelectState } from "react-stately";
 import { useSelect, useButton, mergeProps, useFocusRing } from "react-aria";
 
-import Flex, { FlexProps } from "../SharedComponents/Flex";
 import Icon, { IconName } from "../Icon";
 import getColorByLocation from "../../styles/themeHelpers/getColorByLocation";
 import UnstyledButton from "../UnstyledButton";
@@ -17,6 +16,8 @@ import { Span } from "../Typography";
 
 import { ListBox } from "./ListBox";
 import { Popover } from "./Popover";
+
+import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 
 export { Item } from "react-stately";
