+++ src/components/ErrorView/ErrorView.tsx
@@ -3,11 +3,11 @@ import styled from "styled-components";
 
 import { DEFAULT_SEO_PROPS } from "../../browser-lib/seo/Seo";
 import ButtonAsLink from "../SharedComponents/Button/ButtonAsLink";
-import ButtonGroup from "../ButtonGroup/ButtonGroup";
 import Layout from "../Layout";
 import { HeaderVariant, FooterVariant } from "../Layout/Layout";
 import { P, Heading } from "../Typography";
 
+import ButtonGroup from "@/components/SharedComponents/ButtonGroup";
 import Button from "@/components/SharedComponents/Button";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import Flex from "@/components/SharedComponents/Flex";
