+++ src/components/SharedComponents/ButtonGroup/ButtonGroup.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import styled from "styled-components";
 
-import Flex, { FlexProps } from "../SharedComponents/Flex";
+import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 
 const Group = styled(Flex)`
   & > *:not(:last-child) {
