+++ src/components/SharedComponents/ButtonGroup/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./ButtonGroup";
+export type { ButtonGroupProps } from "./ButtonGroup";
