+++ src/components/SocialButtons/SocialButtons.tsx
@@ -4,8 +4,9 @@ import { PixelSpacing } from "../../styles/theme";
 import { ResponsiveValues } from "../../styles/utils/responsive";
 import { ButtonSize } from "../SharedComponents/Button/common";
 import IconButtonAsLink from "../SharedComponents/Button/IconButtonAsLink";
-import Flex, { FlexProps } from "../SharedComponents/Flex";
 import { IconName } from "../Icon";
+
+import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 
 export const OAK_SOCIALS: Record<SocialNetwork, string> = {
   instagram: "oaknational",
