+++ src/components/TabularNav/TabularNav.tsx
@@ -1,7 +1,8 @@
 import ButtonAsLink, {
   ButtonAsLinkProps,
 } from "../SharedComponents/Button/ButtonAsLink";
-import Flex, { FlexProps } from "../SharedComponents/Flex";
+
+import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 
 /**
  * TabularNav is a 'nav' component which renders 'minimal' (text-link) link
